refactor(home): extract HomeScreen as container and simplify BookListScreen (BYU-193)
- HomeScreen이 모드 전환과 AppBar를 담당하는 컨테이너 역할
- BookListScreen을 순수 목록 화면으로 단순화 (1171줄 → ~200줄)
- PressableWrapper, BookListCard, BookListSkeleton 등 재사용 가능한 위젯 추출
- HomeViewModel이 displayMode 상태 관리, BookListViewModel과 연동
- ReadingProgressScreen으로 진행 중인 독서 상세 화면 분리
- Consumer2로 HomeViewModel과 BookListViewModel 동시 구독하여 데이터 동기화 문제 해결
