+++ app/lib/ui/book_list/widgets/book_list_screen.dart
@@ -1,14 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
-import 'package:shimmer/shimmer.dart';
 
-import 'package:book_golas/domain/models/book.dart';
-import 'package:book_golas/domain/models/home_display_mode.dart';
-import 'package:book_golas/ui/core/widgets/book_image_widget.dart';
 import 'package:book_golas/ui/book_detail/book_detail_screen.dart';
 import 'package:book_golas/ui/book_list/view_model/book_list_view_model.dart';
-import 'package:book_golas/ui/book_list/widgets/sheets/reading_books_selection_sheet.dart';
+import 'package:book_golas/ui/book_list/widgets/book_list_card.dart';
+import 'package:book_golas/ui/book_list/widgets/book_list_skeleton.dart';
 
 class BookListScreen extends StatefulWidget {
   const BookListScreen({super.key});
@@ -21,7 +17,6 @@ class _BookListScreenState extends State<BookListScreen>
     with SingleTickerProviderStateMixin {
   late TabController _tabController;
   bool _isRefreshing = false;
-  final GlobalKey<_PressableEditButtonState> _editButtonKey = GlobalKey();
 
   @override
   void initState() {
@@ -68,198 +63,64 @@ class _BookListScreenState extends State<BookListScreen>
     }
   }
 
-  void _toggleDisplayMode(BookListViewModel vm) {
-    if (vm.displayMode == HomeDisplayMode.allBooks) {
-      vm.setDisplayMode(HomeDisplayMode.readingDetail);
-      _handleReadingDetailMode(vm);
-    } else {
-      vm.setDisplayMode(HomeDisplayMode.allBooks);
-      _showModeChangeSnackBar('전체 독서 보기로 전환되었습니다.');
-    }
-  }
-
-  void _showModeChangeSnackBar(String message) {
-    _editButtonKey.currentState?.triggerTransitionAnimation();
-    ScaffoldMessenger.of(context).hideCurrentSnackBar();
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Row(
-          children: [
-            const Icon(
-              Icons.check_circle_outline,
-              color: Colors.white,
-              size: 20,
-            ),
-            const SizedBox(width: 12),
-            Text(
-              message,
-              style: const TextStyle(
-                fontSize: 14,
-                fontWeight: FontWeight.w500,
-                color: Colors.white,
-              ),
-            ),
-          ],
-        ),
-        behavior: SnackBarBehavior.floating,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(12),
-        ),
-        backgroundColor: const Color(0xFF323232),
-        duration: const Duration(seconds: 2),
-        margin: const EdgeInsets.fromLTRB(16, 0, 16, 24),
-      ),
-    );
-  }
-
-  void _handleReadingDetailMode(BookListViewModel vm) {
-    final readingBooks = vm.readingBooks;
-
-    if (readingBooks.isEmpty) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: Text('진행 중인 독서가 없습니다'),
-          duration: Duration(seconds: 2),
-        ),
-      );
-      vm.setDisplayMode(HomeDisplayMode.allBooks);
-      return;
-    }
-
-    if (readingBooks.length == 1) {
-      vm.setSelectedBook(readingBooks.first.id!);
-      _showModeChangeSnackBar('진행 중인 독서 보기로 전환되었습니다.');
-      return;
-    }
-
-    showModalBottomSheet(
-      context: context,
-      isScrollControlled: true,
-      backgroundColor: Colors.transparent,
-      builder: (_) => ReadingBooksSelectionSheet(
-        books: readingBooks,
-        onBookSelected: (book) {
-          Navigator.pop(context);
-          vm.setSelectedBook(book.id!);
-          _showModeChangeSnackBar('진행 중인 독서 보기로 전환되었습니다.');
-        },
-      ),
-    );
-  }
-
-  String _getToggleButtonLabel(HomeDisplayMode mode) {
-    return mode == HomeDisplayMode.readingDetail ? '전체 독서 보기' : '진행 중인 독서만 보기';
-  }
-
   @override
   Widget build(BuildContext context) {
     final isDark = Theme.of(context).brightness == Brightness.dark;
 
     return Consumer<BookListViewModel>(
       builder: (context, vm, _) {
-        if (!vm.isPreferencesLoaded) {
-          return Scaffold(
-            backgroundColor:
-                isDark ? const Color(0xFF121212) : const Color(0xFFF8F9FA),
-            body: const SizedBox.shrink(),
-          );
-        }
-
-        final isReadingDetailMode =
-            vm.displayMode == HomeDisplayMode.readingDetail;
-        final isReadingDetail = isReadingDetailMode && vm.selectedBook != null;
-
-        if (isReadingDetailMode && vm.selectedBook == null && vm.isLoading) {
-          return Scaffold(
-            backgroundColor:
-                isDark ? const Color(0xFF121212) : const Color(0xFFF8F9FA),
-            body: const SizedBox.shrink(),
-          );
-        }
-
-        return Scaffold(
-          backgroundColor:
-              isDark ? const Color(0xFF121212) : const Color(0xFFF8F9FA),
-          appBar: AppBar(
-            backgroundColor: isReadingDetail ? Colors.transparent : null,
-            elevation: 0,
-            surfaceTintColor: Colors.transparent,
-            scrolledUnderElevation: 0,
-            title: isReadingDetail ? null : const Text('독서 목록'),
-            centerTitle: false,
-            titleTextStyle: TextStyle(
-              fontSize: 20,
-              fontWeight: FontWeight.w600,
-              color: isDark ? Colors.white : Colors.black,
+        return Column(
+          children: [
+            _buildTabBar(vm, isDark),
+            Expanded(
+              child: _buildBody(vm, isDark),
             ),
-            actions: [
-              _PressableEditButton(
-                buttonKey: _editButtonKey,
-                onTap: () => _toggleDisplayMode(vm),
-                isDark: isDark,
-                label: _getToggleButtonLabel(vm.displayMode),
-                icon: Icons.sync_alt,
-              ),
-            ],
-            bottom: isReadingDetail ? null : _buildTabBar(vm, isDark),
-          ),
-          body: isReadingDetail
-              ? BookDetailScreen(
-                  key: ValueKey(vm.selectedBookId),
-                  book: vm.selectedBook!,
-                  isEmbedded: true,
-                )
-              : _buildBody(vm, isDark),
+          ],
         );
       },
     );
   }
 
-  PreferredSize _buildTabBar(BookListViewModel vm, bool isDark) {
+  Widget _buildTabBar(BookListViewModel vm, bool isDark) {
     final selectedTabIndex = vm.selectedTabIndex;
-    return PreferredSize(
-      preferredSize: const Size.fromHeight(50),
-      child: Container(
-        color: isDark ? const Color(0xFF121212) : Colors.white,
-        child: SizedBox(
-          height: 50,
-          child: SingleChildScrollView(
-            scrollDirection: Axis.horizontal,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                SizedBox(
-                  height: 48,
-                  child: Row(
-                    children: [
-                      _buildScrollableTabItem(
-                          '전체', 0, selectedTabIndex, isDark),
-                      _buildScrollableTabItem(
-                          '독서 중', 1, selectedTabIndex, isDark),
-                      _buildScrollableTabItem(
-                          '완독', 2, selectedTabIndex, isDark),
-                      _buildScrollableTabItem(
-                          '다시 읽을 책', 3, selectedTabIndex, isDark),
-                    ],
-                  ),
+    return Container(
+      color: isDark ? const Color(0xFF121212) : Colors.white,
+      child: SizedBox(
+        height: 50,
+        child: SingleChildScrollView(
+          scrollDirection: Axis.horizontal,
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              SizedBox(
+                height: 48,
+                child: Row(
+                  children: [
+                    _buildScrollableTabItem('전체', 0, selectedTabIndex, isDark),
+                    _buildScrollableTabItem(
+                        '독서 중', 1, selectedTabIndex, isDark),
+                    _buildScrollableTabItem('완독', 2, selectedTabIndex, isDark),
+                    _buildScrollableTabItem(
+                        '다시 읽을 책', 3, selectedTabIndex, isDark),
+                  ],
                 ),
-                AnimatedBuilder(
-                  animation: _tabController.animation!,
-                  builder: (context, child) {
-                    const tabWidth = 100.0;
-                    final animationValue = _tabController.animation!.value;
-                    return Transform.translate(
-                      offset: Offset(tabWidth * animationValue, 0),
-                      child: Container(
-                        width: tabWidth,
-                        height: 2,
-                        color: isDark ? Colors.white : Colors.black,
-                      ),
-                    );
-                  },
-                ),
-              ],
-            ),
+              ),
+              AnimatedBuilder(
+                animation: _tabController.animation!,
+                builder: (context, child) {
+                  const tabWidth = 100.0;
+                  final animationValue = _tabController.animation!.value;
+                  return Transform.translate(
+                    offset: Offset(tabWidth * animationValue, 0),
+                    child: Container(
+                      width: tabWidth,
+                      height: 2,
+                      color: isDark ? Colors.white : Colors.black,
+                    ),
+                  );
+                },
+              ),
+            ],
           ),
         ),
       ),
@@ -268,55 +129,11 @@ class _BookListScreenState extends State<BookListScreen>
 
   Widget _buildBody(BookListViewModel vm, bool isDark) {
     if (vm.isLoading && vm.books.isEmpty) {
-      return _buildSkeletonList(isDark);
+      return const BookListSkeleton();
     }
 
     if (vm.errorMessage != null) {
-      return Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(
-              Icons.error_outline,
-              size: 60,
-              color: isDark ? Colors.grey[400] : Colors.grey[600],
-            ),
-            const SizedBox(height: 16),
-            Text(
-              '데이터를 불러올 수 없습니다',
-              style: TextStyle(
-                fontSize: 18,
-                fontWeight: FontWeight.w500,
-                color: isDark ? Colors.white : Colors.black,
-              ),
-            ),
-            const SizedBox(height: 8),
-            Text(
-              '네트워크 연결을 확인해주세요',
-              style: TextStyle(
-                fontSize: 14,
-                color: isDark ? Colors.grey[400] : Colors.grey[600],
-              ),
-            ),
-            const SizedBox(height: 24),
-            ElevatedButton.icon(
-              onPressed: () {
-                setState(() {});
-              },
-              icon: const Icon(Icons.refresh),
-              label: const Text('다시 시도'),
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.blue,
-                foregroundColor: Colors.white,
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 24,
-                  vertical: 12,
-                ),
-              ),
-            ),
-          ],
-        ),
-      );
+      return _buildErrorState(isDark);
     }
 
     return TabBarView(
@@ -355,6 +172,54 @@ class _BookListScreenState extends State<BookListScreen>
     );
   }
 
+  Widget _buildErrorState(bool isDark) {
+    return Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Icon(
+            Icons.error_outline,
+            size: 60,
+            color: isDark ? Colors.grey[400] : Colors.grey[600],
+          ),
+          const SizedBox(height: 16),
+          Text(
+            '데이터를 불러올 수 없습니다',
+            style: TextStyle(
+              fontSize: 18,
+              fontWeight: FontWeight.w500,
+              color: isDark ? Colors.white : Colors.black,
+            ),
+          ),
+          const SizedBox(height: 8),
+          Text(
+            '네트워크 연결을 확인해주세요',
+            style: TextStyle(
+              fontSize: 14,
+              color: isDark ? Colors.grey[400] : Colors.grey[600],
+            ),
+          ),
+          const SizedBox(height: 24),
+          ElevatedButton.icon(
+            onPressed: () {
+              setState(() {});
+            },
+            icon: const Icon(Icons.refresh),
+            label: const Text('다시 시도'),
+            style: ElevatedButton.styleFrom(
+              backgroundColor: Colors.blue,
+              foregroundColor: Colors.white,
+              padding: const EdgeInsets.symmetric(
+                horizontal: 24,
+                vertical: 12,
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
   Widget _buildRereadBooksTab(BookListViewModel vm, bool isDark) {
     return Center(
       child: Column(
@@ -387,66 +252,20 @@ class _BookListScreenState extends State<BookListScreen>
     );
   }
 
-  Widget _buildReadingEmptyState() {
-    return const Center(
+  Widget _buildEmptyState(String message) {
+    return Center(
       child: Column(
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
-          Icon(
+          const Icon(
             Icons.book_outlined,
             size: 80,
             color: Colors.grey,
           ),
-          SizedBox(height: 16),
+          const SizedBox(height: 16),
           Text(
-            '현재 읽고 있는 책이 없습니다',
-            style: TextStyle(
-              fontSize: 18,
-              color: Colors.grey,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildCompletedEmptyState() {
-    return const Center(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Icon(
-            Icons.book_outlined,
-            size: 80,
-            color: Colors.grey,
-          ),
-          SizedBox(height: 16),
-          Text(
-            '완독한 책이 없습니다',
-            style: TextStyle(
-              fontSize: 18,
-              color: Colors.grey,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildEmptyState() {
-    return const Center(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Icon(
-            Icons.book_outlined,
-            size: 80,
-            color: Colors.grey,
-          ),
-          SizedBox(height: 16),
-          Text(
-            '아직 시작한 독서가 없습니다',
-            style: TextStyle(
+            message,
+            style: const TextStyle(
               fontSize: 18,
               color: Colors.grey,
             ),
@@ -461,7 +280,7 @@ class _BookListScreenState extends State<BookListScreen>
     final readingBooks = vm.readingBooks;
 
     if (allBooks.isEmpty) {
-      return _buildEmptyState();
+      return _buildEmptyState('아직 시작한 독서가 없습니다');
     }
 
     return RefreshIndicator(
@@ -500,7 +319,10 @@ class _BookListScreenState extends State<BookListScreen>
             ),
             const SizedBox(height: 12),
             ...(vm.showAllCurrentBooks ? readingBooks : readingBooks.take(3))
-                .map((book) => _buildBookCard(book)),
+                .map((book) => BookListCard(
+                      book: book,
+                      onTap: () => _navigateToBookDetail(book),
+                    )),
             const SizedBox(height: 24),
             Divider(color: isDark ? Colors.grey[700] : Colors.grey[300]),
             const SizedBox(height: 24),
@@ -514,7 +336,10 @@ class _BookListScreenState extends State<BookListScreen>
             ),
           ),
           const SizedBox(height: 12),
-          ...allBooks.map((book) => _buildBookCard(book)),
+          ...allBooks.map((book) => BookListCard(
+                book: book,
+                onTap: () => _navigateToBookDetail(book),
+              )),
         ],
       ),
     );
@@ -524,7 +349,7 @@ class _BookListScreenState extends State<BookListScreen>
     final readingBooks = vm.readingBooks;
 
     if (readingBooks.isEmpty) {
-      return _buildReadingEmptyState();
+      return _buildEmptyState('현재 읽고 있는 책이 없습니다');
     }
 
     return RefreshIndicator(
@@ -534,7 +359,12 @@ class _BookListScreenState extends State<BookListScreen>
       child: ListView(
         physics: const AlwaysScrollableScrollPhysics(),
         padding: const EdgeInsets.fromLTRB(16, 16, 16, 200),
-        children: readingBooks.map((book) => _buildBookCard(book)).toList(),
+        children: readingBooks
+            .map((book) => BookListCard(
+                  book: book,
+                  onTap: () => _navigateToBookDetail(book),
+                ))
+            .toList(),
       ),
     );
   }
@@ -543,7 +373,7 @@ class _BookListScreenState extends State<BookListScreen>
     final completedBooks = vm.completedBooks;
 
     if (completedBooks.isEmpty) {
-      return _buildCompletedEmptyState();
+      return _buildEmptyState('완독한 책이 없습니다');
     }
 
     return RefreshIndicator(
@@ -553,618 +383,25 @@ class _BookListScreenState extends State<BookListScreen>
       child: ListView(
         physics: const AlwaysScrollableScrollPhysics(),
         padding: const EdgeInsets.fromLTRB(16, 16, 16, 200),
-        children: completedBooks.map((book) => _buildBookCard(book)).toList(),
+        children: completedBooks
+            .map((book) => BookListCard(
+                  book: book,
+                  onTap: () => _navigateToBookDetail(book),
+                ))
+            .toList(),
       ),
     );
   }
 
-  Widget _buildBookCard(Book book) {
-    final isDark = Theme.of(context).brightness == Brightness.dark;
-
-    return _PressableBookCard(
-      book: book,
-      isDark: isDark,
-      onTap: () async {
-        await Navigator.push(
-          context,
-          MaterialPageRoute(
-            builder: (context) => BookDetailScreen(book: book),
-          ),
-        );
-        if (mounted) {
-          context.read<BookListViewModel>().refresh();
-        }
-      },
-    );
-  }
-
-  Widget _buildSkeletonList(bool isDark) {
-    return ListView.builder(
-      padding: const EdgeInsets.fromLTRB(16, 16, 16, 200),
-      itemCount: 3,
-      itemBuilder: (context, index) => _buildSkeletonCard(isDark, index),
-    );
-  }
-
-  Widget _buildSkeletonCard(bool isDark, [int index = 0]) {
-    final titleWidths = [double.infinity, 180.0, 220.0];
-    final subtitleWidths = [140.0, 100.0, 160.0];
-
-    return Shimmer.fromColors(
-      baseColor: isDark ? Colors.grey[800]! : Colors.grey[300]!,
-      highlightColor: isDark ? Colors.grey[700]! : Colors.grey[100]!,
-      child: Container(
-        margin: const EdgeInsets.only(bottom: 16),
-        padding: const EdgeInsets.all(16),
-        decoration: BoxDecoration(
-          color: isDark ? const Color(0xFF1E1E1E) : Colors.white,
-          borderRadius: BorderRadius.circular(8),
-          boxShadow: [
-            BoxShadow(
-              color: isDark
-                  ? Colors.black.withValues(alpha: 0.3)
-                  : Colors.grey.withValues(alpha: 0.1),
-              spreadRadius: 1,
-              blurRadius: 3,
-              offset: const Offset(0, 1),
-            ),
-          ],
-        ),
-        child: Row(
-          children: [
-            Container(
-              width: 60,
-              height: 80,
-              decoration: BoxDecoration(
-                color: isDark ? Colors.grey[700] : Colors.grey[300],
-                borderRadius: BorderRadius.circular(4),
-                border: Border.all(
-                  color: isDark ? Colors.grey[600]! : Colors.grey[400]!,
-                ),
-              ),
-            ),
-            const SizedBox(width: 16),
-            Expanded(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Container(
-                    height: 16,
-                    width: titleWidths[index % 3],
-                    decoration: BoxDecoration(
-                      color: isDark ? Colors.grey[700] : Colors.grey[300],
-                      borderRadius: BorderRadius.circular(4),
-                    ),
-                  ),
-                  const SizedBox(height: 6),
-                  Row(
-                    children: [
-                      Container(
-                        height: 24,
-                        width: 52,
-                        decoration: BoxDecoration(
-                          color: isDark ? Colors.grey[700] : Colors.grey[300],
-                          borderRadius: BorderRadius.circular(6),
-                        ),
-                      ),
-                      const SizedBox(width: 8),
-                      Container(
-                        height: 14,
-                        width: subtitleWidths[index % 3],
-                        decoration: BoxDecoration(
-                          color: isDark ? Colors.grey[700] : Colors.grey[300],
-                          borderRadius: BorderRadius.circular(4),
-                        ),
-                      ),
-                    ],
-                  ),
-                  const SizedBox(height: 8),
-                  Row(
-                    children: [
-                      Expanded(
-                        child: Container(
-                          height: 6,
-                          decoration: BoxDecoration(
-                            color: isDark ? Colors.grey[700] : Colors.grey[300],
-                            borderRadius: BorderRadius.circular(4),
-                          ),
-                        ),
-                      ),
-                      const SizedBox(width: 8),
-                      Container(
-                        height: 12,
-                        width: 32,
-                        decoration: BoxDecoration(
-                          color: isDark ? Colors.grey[700] : Colors.grey[300],
-                          borderRadius: BorderRadius.circular(4),
-                        ),
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-            Icon(
-              Icons.arrow_forward_ios,
-              color: isDark ? Colors.grey[700] : Colors.grey[300],
-              size: 16,
-            ),
-          ],
-        ),
+  void _navigateToBookDetail(book) async {
+    await Navigator.push(
+      context,
+      MaterialPageRoute(
+        builder: (context) => BookDetailScreen(book: book),
       ),
     );
-  }
-}
-
-class _PressableBookCard extends StatefulWidget {
-  final Book book;
-  final bool isDark;
-  final VoidCallback onTap;
-
-  const _PressableBookCard({
-    required this.book,
-    required this.isDark,
-    required this.onTap,
-  });
-
-  @override
-  State<_PressableBookCard> createState() => _PressableBookCardState();
-}
-
-class _PressableBookCardState extends State<_PressableBookCard>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _controller;
-  late Animation<double> _scaleAnimation;
-  late Animation<double> _brightnessAnimation;
-  final GlobalKey _cardKey = GlobalKey();
-  bool _isLongPressing = false;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = AnimationController(
-      duration: const Duration(milliseconds: 150),
-      vsync: this,
-    );
-    _scaleAnimation = Tween<double>(begin: 1.0, end: 0.96).animate(
-      CurvedAnimation(parent: _controller, curve: Curves.easeInOut),
-    );
-    _brightnessAnimation = Tween<double>(begin: 0.0, end: 0.1).animate(
-      CurvedAnimation(parent: _controller, curve: Curves.easeInOut),
-    );
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  void _onTapDown(TapDownDetails details) {
-    _controller.forward();
-  }
-
-  void _onTapUp(TapUpDetails details) {
-    Future.delayed(const Duration(milliseconds: 80), () {
-      if (mounted) {
-        _controller.reverse().then((_) {
-          if (mounted) {
-            widget.onTap();
-          }
-        });
-      }
-    });
-  }
-
-  void _onTapCancel() {
-    // onTapCancel이 onLongPressStart보다 먼저 호출되므로
-    // 잠시 대기 후 _isLongPressing 체크
-    Future.delayed(const Duration(milliseconds: 50), () {
-      if (!_isLongPressing && mounted) {
-        _controller.reverse();
-      }
-    });
-  }
-
-  void _onLongPressStart(LongPressStartDetails details) {
-    _isLongPressing = true;
-    HapticFeedback.mediumImpact();
-  }
-
-  void _onLongPressEnd(LongPressEndDetails details) {
-    _isLongPressing = false;
-    _controller.reverse();
-
-    final RenderBox? renderBox =
-        _cardKey.currentContext?.findRenderObject() as RenderBox?;
-    if (renderBox != null) {
-      final localPosition = renderBox.globalToLocal(details.globalPosition);
-      final isInsideCard = localPosition.dx >= 0 &&
-          localPosition.dx <= renderBox.size.width &&
-          localPosition.dy >= 0 &&
-          localPosition.dy <= renderBox.size.height;
-
-      if (isInsideCard) {
-        HapticFeedback.lightImpact();
-        widget.onTap();
-      }
+    if (mounted) {
+      context.read<BookListViewModel>().refresh();
     }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final book = widget.book;
-    final isDark = widget.isDark;
-
-    final now = DateTime.now();
-    final today = DateTime(now.year, now.month, now.day);
-    final target = DateTime(
-        book.targetDate.year, book.targetDate.month, book.targetDate.day);
-    final daysLeft = target.difference(today).inDays;
-    final pageProgress = book.totalPages > 0
-        ? (book.currentPage / book.totalPages).clamp(0.0, 1.0)
-        : 0.0;
-    final isCompleted =
-        book.currentPage >= book.totalPages && book.totalPages > 0;
-
-    return GestureDetector(
-      key: _cardKey,
-      onTapDown: _onTapDown,
-      onTapUp: _onTapUp,
-      onTapCancel: _onTapCancel,
-      onLongPressStart: _onLongPressStart,
-      onLongPressEnd: _onLongPressEnd,
-      child: AnimatedBuilder(
-        animation: _controller,
-        builder: (context, child) {
-          return Transform.scale(
-            scale: _scaleAnimation.value,
-            child: Container(
-              margin: const EdgeInsets.only(bottom: 16),
-              foregroundDecoration: BoxDecoration(
-                color:
-                    Colors.white.withValues(alpha: _brightnessAnimation.value),
-                borderRadius: BorderRadius.circular(8),
-              ),
-              decoration: BoxDecoration(
-                color: isDark ? const Color(0xFF1E1E1E) : Colors.white,
-                borderRadius: BorderRadius.circular(8),
-                boxShadow: [
-                  BoxShadow(
-                    color: isDark
-                        ? Colors.black.withValues(alpha: 0.3)
-                        : Colors.grey.withValues(alpha: 0.1),
-                    spreadRadius: 1,
-                    blurRadius: 3,
-                    offset: const Offset(0, 1),
-                  ),
-                ],
-              ),
-              child: Padding(
-                padding: const EdgeInsets.all(16),
-                child: Row(
-                  children: [
-                    Container(
-                      width: 60,
-                      height: 80,
-                      decoration: BoxDecoration(
-                        border: Border.all(
-                          color: isDark ? Colors.grey[700]! : Colors.grey[300]!,
-                        ),
-                        borderRadius: BorderRadius.circular(4),
-                      ),
-                      child: ClipRRect(
-                        borderRadius: BorderRadius.circular(4),
-                        child: BookImageWidget(
-                          imageUrl: book.imageUrl,
-                          iconSize: 30,
-                        ),
-                      ),
-                    ),
-                    const SizedBox(width: 16),
-                    Expanded(
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          Text(
-                            book.title,
-                            style: TextStyle(
-                              fontSize: 16,
-                              fontWeight: FontWeight.bold,
-                              color: isDark ? Colors.white : Colors.black,
-                            ),
-                            maxLines: 2,
-                            overflow: TextOverflow.ellipsis,
-                          ),
-                          const SizedBox(height: 6),
-                          Row(
-                            children: [
-                              Container(
-                                padding: const EdgeInsets.symmetric(
-                                    horizontal: 8, vertical: 4),
-                                decoration: BoxDecoration(
-                                  color: daysLeft < 0
-                                      ? const Color(0xFFEF4444)
-                                          .withValues(alpha: 0.12)
-                                      : (isCompleted
-                                          ? const Color(0xFF10B981)
-                                              .withValues(alpha: 0.12)
-                                          : const Color(0xFF5B7FFF)
-                                              .withValues(alpha: 0.12)),
-                                  borderRadius: BorderRadius.circular(6),
-                                ),
-                                child: Text(
-                                  daysLeft >= 0
-                                      ? 'D-$daysLeft'
-                                      : 'D+${daysLeft.abs()}',
-                                  style: TextStyle(
-                                    fontSize: 12,
-                                    fontWeight: FontWeight.w600,
-                                    color: daysLeft < 0
-                                        ? const Color(0xFFEF4444)
-                                        : (isCompleted
-                                            ? const Color(0xFF10B981)
-                                            : const Color(0xFF5B7FFF)),
-                                  ),
-                                ),
-                              ),
-                              const SizedBox(width: 8),
-                              Text(
-                                '${book.currentPage}/${book.totalPages}페이지',
-                                style: TextStyle(
-                                  fontSize: 13,
-                                  color: isDark
-                                      ? Colors.grey[400]
-                                      : Colors.grey[600],
-                                ),
-                              ),
-                            ],
-                          ),
-                          const SizedBox(height: 8),
-                          Row(
-                            children: [
-                              Expanded(
-                                child: ClipRRect(
-                                  borderRadius: BorderRadius.circular(4),
-                                  child: LinearProgressIndicator(
-                                    value: pageProgress,
-                                    backgroundColor: isDark
-                                        ? Colors.grey[700]
-                                        : Colors.grey[200],
-                                    valueColor: AlwaysStoppedAnimation<Color>(
-                                      isCompleted
-                                          ? const Color(0xFF10B981)
-                                          : const Color(0xFF5B7FFF),
-                                    ),
-                                    minHeight: 6,
-                                  ),
-                                ),
-                              ),
-                              const SizedBox(width: 8),
-                              Text(
-                                '${(pageProgress * 100).toStringAsFixed(0)}%',
-                                style: TextStyle(
-                                  fontSize: 12,
-                                  fontWeight: FontWeight.w600,
-                                  color: isCompleted
-                                      ? const Color(0xFF10B981)
-                                      : const Color(0xFF5B7FFF),
-                                ),
-                              ),
-                            ],
-                          ),
-                        ],
-                      ),
-                    ),
-                    Icon(
-                      Icons.arrow_forward_ios,
-                      color: isDark ? Colors.grey[400] : Colors.grey,
-                      size: 16,
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          );
-        },
-      ),
-    );
-  }
-}
-
-class _PressableEditButton extends StatefulWidget {
-  final GlobalKey buttonKey;
-  final VoidCallback onTap;
-  final bool isDark;
-  final String label;
-  final IconData icon;
-
-  const _PressableEditButton({
-    required this.buttonKey,
-    required this.onTap,
-    required this.isDark,
-    required this.label,
-    required this.icon,
-  });
-
-  @override
-  State<_PressableEditButton> createState() => _PressableEditButtonState();
-}
-
-class _PressableEditButtonState extends State<_PressableEditButton>
-    with TickerProviderStateMixin {
-  late AnimationController _pressController;
-  late Animation<double> _scaleAnimation;
-  late Animation<double> _brightnessAnimation;
-  bool _isLongPressing = false;
-
-  late AnimationController _transitionController;
-  late Animation<double> _rotationAnimation;
-  late Animation<double> _fadeOutAnimation;
-  late Animation<double> _fadeInAnimation;
-  String _displayedLabel = '';
-
-  @override
-  void initState() {
-    super.initState();
-    _displayedLabel = widget.label;
-
-    _pressController = AnimationController(
-      duration: const Duration(milliseconds: 150),
-      vsync: this,
-    );
-    _scaleAnimation = Tween<double>(begin: 1.0, end: 0.96).animate(
-      CurvedAnimation(parent: _pressController, curve: Curves.easeInOut),
-    );
-    _brightnessAnimation = Tween<double>(begin: 0.0, end: 0.1).animate(
-      CurvedAnimation(parent: _pressController, curve: Curves.easeInOut),
-    );
-
-    _transitionController = AnimationController(
-      duration: const Duration(milliseconds: 400),
-      vsync: this,
-    );
-    _rotationAnimation = Tween<double>(begin: 0.0, end: 1.0).animate(
-      CurvedAnimation(parent: _transitionController, curve: Curves.easeInOut),
-    );
-    _fadeOutAnimation = Tween<double>(begin: 1.0, end: 0.0).animate(
-      CurvedAnimation(
-        parent: _transitionController,
-        curve: const Interval(0.0, 0.5, curve: Curves.easeOut),
-      ),
-    );
-    _fadeInAnimation = Tween<double>(begin: 0.0, end: 1.0).animate(
-      CurvedAnimation(
-        parent: _transitionController,
-        curve: const Interval(0.5, 1.0, curve: Curves.easeIn),
-      ),
-    );
-
-    _transitionController.addListener(() {
-      if (_transitionController.value >= 0.5 &&
-          _displayedLabel != widget.label) {
-        setState(() {
-          _displayedLabel = widget.label;
-        });
-      }
-    });
-  }
-
-  @override
-  void didUpdateWidget(covariant _PressableEditButton oldWidget) {
-    super.didUpdateWidget(oldWidget);
-  }
-
-  void triggerTransitionAnimation() {
-    _transitionController.forward(from: 0.0);
-  }
-
-  @override
-  void dispose() {
-    _pressController.dispose();
-    _transitionController.dispose();
-    super.dispose();
-  }
-
-  void _onTapDown(TapDownDetails details) {
-    _pressController.forward();
-  }
-
-  void _onTapUp(TapUpDetails details) {
-    Future.delayed(const Duration(milliseconds: 80), () {
-      if (mounted) {
-        _pressController.reverse().then((_) {
-          if (mounted) {
-            widget.onTap();
-          }
-        });
-      }
-    });
-  }
-
-  void _onTapCancel() {
-    Future.delayed(const Duration(milliseconds: 50), () {
-      if (!_isLongPressing && mounted) {
-        _pressController.reverse();
-      }
-    });
-  }
-
-  void _onLongPressStart(LongPressStartDetails details) {
-    _isLongPressing = true;
-    HapticFeedback.mediumImpact();
-  }
-
-  void _onLongPressEnd(LongPressEndDetails details) {
-    _isLongPressing = false;
-    _pressController.reverse();
-    HapticFeedback.lightImpact();
-    widget.onTap();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      key: widget.buttonKey,
-      onTapDown: _onTapDown,
-      onTapUp: _onTapUp,
-      onTapCancel: _onTapCancel,
-      onLongPressStart: _onLongPressStart,
-      onLongPressEnd: _onLongPressEnd,
-      child: AnimatedBuilder(
-        animation: _pressController,
-        builder: (context, child) {
-          return Transform.scale(
-            scale: _scaleAnimation.value,
-            child: Container(
-              padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-              foregroundDecoration: BoxDecoration(
-                color:
-                    Colors.white.withValues(alpha: _brightnessAnimation.value),
-                borderRadius: BorderRadius.circular(8),
-              ),
-              child: Row(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  AnimatedBuilder(
-                    animation: _transitionController,
-                    builder: (context, child) {
-                      return Transform.rotate(
-                        angle: _rotationAnimation.value * 2 * 3.14159,
-                        child: Icon(
-                          widget.icon,
-                          size: 18,
-                          color: widget.isDark ? Colors.white : Colors.black,
-                        ),
-                      );
-                    },
-                  ),
-                  const SizedBox(width: 6),
-                  AnimatedBuilder(
-                    animation: _transitionController,
-                    builder: (context, child) {
-                      final opacity = _transitionController.value <= 0.5
-                          ? _fadeOutAnimation.value
-                          : _fadeInAnimation.value;
-                      return Opacity(
-                        opacity: opacity,
-                        child: Text(
-                          _displayedLabel,
-                          style: TextStyle(
-                            fontSize: 14,
-                            fontWeight: FontWeight.w500,
-                            color: widget.isDark ? Colors.white : Colors.black,
-                          ),
-                        ),
-                      );
-                    },
-                  ),
-                ],
-              ),
-            ),
-          );
-        },
-      ),
-    );
   }
 }
